- move MainApp to file.
