+++ lib/main.dart
@@ -1,20 +1,7 @@
 import 'package:flutter/material.dart';
 
+import 'main_app.dart';
+
 void main() {
   runApp(const MainApp());
-}
-
-class MainApp extends StatelessWidget {
-  const MainApp({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return const MaterialApp(
-      home: Scaffold(
-        body: Center(
-          child: Text('Hello World!'),
-        ),
-      ),
-    );
-  }
 }
